Move Social.ts in a social directory
